+++ src/test/java/com/github/victools/jsonschema/generator/impl/module/SimpleTypeModuleTest.java
@@ -21,22 +21,32 @@ import com.github.victools.jsonschema.generator.FieldScope;
 import com.github.victools.jsonschema.generator.MethodScope;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfigBuilder;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfigPart;
+import com.github.victools.jsonschema.generator.SchemaGeneratorTypeConfigPart;
+import com.github.victools.jsonschema.generator.TypeScope;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Answers;
+import org.mockito.Mock;
 import org.mockito.Mockito;
+import org.mockito.junit.MockitoJUnitRunner;
 
 /**
  * Test for {@link SimpleTypeModule} class.
  */
+@RunWith(MockitoJUnitRunner.class)
 public class SimpleTypeModuleTest {
 
     private SimpleTypeModule instance;
+    @Mock(answer = Answers.RETURNS_DEEP_STUBS)
     private SchemaGeneratorConfigBuilder builder;
+    @Mock(answer = Answers.RETURNS_SELF)
+    private SchemaGeneratorTypeConfigPart<TypeScope> typeConfigPart;
 
     @Before
     public void setUp() {
         this.instance = new SimpleTypeModule();
-        this.builder = Mockito.mock(SchemaGeneratorConfigBuilder.class, Mockito.RETURNS_DEEP_STUBS);
+        Mockito.when(this.builder.forTypesInGeneral()).thenReturn(this.typeConfigPart);
     }
 
     @Test
@@ -52,6 +62,11 @@ public class SimpleTypeModuleTest {
         Mockito.verify(methodConfigPart).withNullableCheck(Mockito.any());
         Mockito.verifyNoMoreInteractions(methodConfigPart);
         Mockito.verify(this.builder, Mockito.times(2)).forMethods();
+
+        Mockito.verify(this.typeConfigPart).withAdditionalPropertiesResolver(Mockito.any());
+        Mockito.verify(this.typeConfigPart).withPatternPropertiesResolver(Mockito.any());
+        Mockito.verifyNoMoreInteractions(this.typeConfigPart);
+        Mockito.verify(this.builder).forTypesInGeneral();
 
         Mockito.verify(this.builder).getObjectMapper();
         Mockito.verify(this.builder).with(Mockito.any(CustomDefinitionProviderV2.class));