+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [4.1.0] - 2020-02-18
 ### Added
 - new `Option.FLATTENED_ENUMS_FROM_TOSTRING`, using `toString()` instead of `name()` as per `Option.FLATTENED_ENUMS`
@@ -108,6 +110,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Specific handling of optionals (two alternatives via standard options).
 - Pre-defined sets of standard options to cover different use-cases and simplify library usage.
 
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.1.0...HEAD
 [4.1.0]: https://github.com/victools/jsonschema-generator/compare/v4.0.2...v4.1.0
 [4.0.2]: https://github.com/victools/jsonschema-generator/compare/v4.0.1...v4.0.2
 [4.0.1]: https://github.com/victools/jsonschema-generator/compare/v4.0.0...v4.0.1