+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for "additionalProperties" property via `SchemaGeneratorConfigPart.withAdditionalPropertiesResolver()` (#22)
 
 ## [4.1.0] - 2020-02-18
 ### Added